Fixes #1
This also closes tex-persian/xepersian/#12

The cause of the issue was applying left to right mark (U+200E) and I removed it to fix the issue. Applying U+200E was a nonstandard approach and only was specific to XB Series fonts and various other derived other fonts. This worked few years ago but now the behavior of XeTeX / Harfbuzz has changed and this ugly hack no longer works. We should indeed avoid using such nonstandard approaches.

However, this commit reintroduces an old issue with XB Series font with applying U+200D after U+0647:

\font\testfont="[Yas.ttf]:script=arab" at 100pt
\testfont
\char"0647\char"200D\char"0631
\bye

This is a XB Series font issue really and it should be fixed in the font itself.